robo/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
comtypes==1.4.16
cryptography==46.0.6
distro==1.9.0
filelock==3.25.2
fsspec==2026.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-genai==1.70.0
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
groq==0.31.0
grpcio==1.80.0
grpcio-status==1.71.2
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.13.0
langdetect==1.0.9
llvmlite==0.47.0
MarkupSafe==3.0.3
more-itertools==11.0.1
mpmath==1.3.0
networkx==3.6.1
numba==0.65.0
numpy==2.4.4
openai==2.30.0
openai-whisper==20250625
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.3.2
pypiwin32==223
PyQt6==6.11.0
PyQt6-Qt6==6.11.0
PyQt6_sip==13.11.1
python-dotenv==1.2.2
pyttsx3==2.99
pywin32==311
regex==2026.4.4
requests==2.32.5
setuptools==81.0.0
simpleaudio==1.0.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.5
soundfile==0.13.1
SpeechRecognition==3.15.2
sympy==1.14.0
tenacity==9.1.4
tiktoken==0.12.0
torch==2.11.0
tqdm==4.67.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
websockets==16.0
